Added Windows and MCloud middleware
