+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/capability/middleware/CapabilitiesMiddleware.java
@@ -1,21 +1,21 @@
-package com.qaprosoft.carina.core.foundation.webdriver.core.capability.chain;
+package com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware;
 
 import java.lang.invoke.MethodHandles;
 
 import org.openqa.selenium.Capabilities;
+import org.openqa.selenium.MutableCapabilities;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public abstract class CapabilitiesExternalMiddleware {
-
+public abstract class CapabilitiesMiddleware {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
-    private CapabilitiesExternalMiddleware next;
+    private CapabilitiesMiddleware next;
 
-    public static CapabilitiesExternalMiddleware link(CapabilitiesExternalMiddleware first, CapabilitiesExternalMiddleware... chain) {
+    public static CapabilitiesMiddleware link(CapabilitiesMiddleware first, CapabilitiesMiddleware... chain) {
 
-        CapabilitiesExternalMiddleware head = first;
-        for (CapabilitiesExternalMiddleware nextInChain : chain) {
+        CapabilitiesMiddleware head = first;
+        for (CapabilitiesMiddleware nextInChain : chain) {
             head.next = nextInChain;
             head = nextInChain;
         }
@@ -24,16 +24,17 @@ public abstract class CapabilitiesExternalMiddleware {
 
     protected abstract boolean isDetected(Capabilities capabilities);
 
-    protected abstract Capabilities upgradeCapabilities(Capabilities capabilities);
+    protected abstract MutableCapabilities upgradeCapabilities(MutableCapabilities capabilities);
 
-    public Capabilities analyze(Capabilities capabilities) {
+    public MutableCapabilities analyze(MutableCapabilities capabilities) {
+
         if (!isDetected(capabilities)) {
             if (next == null) {
-                LOGGER.debug("Integrations was not detected");
                 return capabilities;
             }
             return next.analyze(capabilities);
         }
+
         return upgradeCapabilities(capabilities);
     }
 }
