+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/capability/middleware/FirefoxCapabilitiesMiddleware.java
@@ -1,23 +1,23 @@
-package com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain;
+package com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware;
 
 import java.util.Objects;
 
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.MutableCapabilities;
-import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.firefox.FirefoxOptions;
 import org.openqa.selenium.remote.Browser;
 
 import com.qaprosoft.carina.core.foundation.commons.SpecialKeywords;
 import com.qaprosoft.carina.core.foundation.utils.Configuration;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.impl.ChromeCapabilities;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.impl.FirefoxCapabilities;
 
-public class ChromeCapabilitiesMiddleware extends CapabilitiesMiddleware {
+public class FirefoxCapabilitiesMiddleware extends CapabilitiesMiddleware {
 
     @Override
     protected boolean isDetected(Capabilities capabilities) {
 
         // it is not nesessary to upgrade capabilities if it is instance of ChromeOptions
-        if (capabilities instanceof ChromeOptions) {
+        if (capabilities instanceof FirefoxOptions) {
             return false;
         }
 
@@ -25,7 +25,7 @@ public class ChromeCapabilitiesMiddleware extends CapabilitiesMiddleware {
             return false;
         }
 
-        if (Browser.CHROME.browserName().equalsIgnoreCase(capabilities.getBrowserName())) {
+        if (Browser.FIREFOX.browserName().equalsIgnoreCase(capabilities.getBrowserName())) {
             return true;
         }
         return false;
@@ -33,6 +33,6 @@ public class ChromeCapabilitiesMiddleware extends CapabilitiesMiddleware {
 
     @Override
     protected MutableCapabilities upgradeCapabilities(MutableCapabilities capabilities) {
-        return new ChromeCapabilities().getCapabilities(capabilities);
+        return new FirefoxCapabilities().getCapabilities(capabilities);
     }
 }
