+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/capability/middleware/WindowsCapabilitiesMiddleware.java
@@ -1,36 +1,34 @@
-package com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain;
+package com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware;
 
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.MutableCapabilities;
 
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.impl.GeckoCapabilities;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.impl.WindowsCapabilities;
 
-import io.appium.java_client.gecko.options.GeckoOptions;
 import io.appium.java_client.remote.AutomationName;
 import io.appium.java_client.remote.options.SupportsAutomationNameOption;
+import io.appium.java_client.windows.options.WindowsOptions;
 
-public class GeckoCapabilitiesMiddleware extends CapabilitiesMiddleware {
+public class WindowsCapabilitiesMiddleware extends CapabilitiesMiddleware {
 
     @Override
     protected boolean isDetected(Capabilities capabilities) {
 
-        // it is not nesessary to upgrade capabilities if it is instance of GeckoOptions
-        if (capabilities instanceof GeckoOptions) {
+        if (capabilities instanceof WindowsOptions) {
             return false;
         }
 
         if (capabilities.getCapability(SupportsAutomationNameOption.AUTOMATION_NAME_OPTION) != null &&
                 capabilities.getCapability(SupportsAutomationNameOption.AUTOMATION_NAME_OPTION)
                         .toString()
-                        .equalsIgnoreCase(AutomationName.GECKO)) {
+                        .equalsIgnoreCase(AutomationName.WINDOWS)) {
             return true;
         }
-
         return false;
     }
 
     @Override
     protected MutableCapabilities upgradeCapabilities(MutableCapabilities capabilities) {
-        return new GeckoCapabilities().getCapabilities(capabilities);
+        return new WindowsCapabilities().getCapabilities(capabilities);
     }
 }
