+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/factory/DriverFactory.java
@@ -31,19 +31,22 @@ import com.qaprosoft.carina.core.foundation.utils.Configuration;
 import com.qaprosoft.carina.core.foundation.utils.Configuration.Parameter;
 import com.qaprosoft.carina.core.foundation.utils.R;
 import com.qaprosoft.carina.core.foundation.webdriver.core.capability.AbstractCapabilities;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.BrowserstackPostMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.BrowserstackPreMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.CapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.ChromeCapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.EdgeCapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.FirefoxCapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.GeckoCapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.Mac2CapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.SafariCapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.SauceLabsPreMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.UIAutomaror2CapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.XCUITestCapabilitiesMiddleware;
-import com.qaprosoft.carina.core.foundation.webdriver.core.capability.capabilchain.ZebrunnerPreMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.BrowserstackPostMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.BrowserstackPreMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.CapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.ChromeCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.EdgeCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.FirefoxCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.GeckoCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.LambdaTestPreMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.MCloudPostMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.Mac2CapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.SafariCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.SauceLabsPreMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.UIAutomaror2CapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.WindowsCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.XCUITestCapabilitiesMiddleware;
+import com.qaprosoft.carina.core.foundation.webdriver.core.capability.middleware.ZebrunnerPreMiddleware;
 import com.qaprosoft.carina.core.foundation.webdriver.core.factory.chain.AbstractFactory;
 import com.qaprosoft.carina.core.foundation.webdriver.core.factory.chain.AndroidFactory;
 import com.qaprosoft.carina.core.foundation.webdriver.core.factory.chain.ChromeFactory;
@@ -79,7 +82,8 @@ public class DriverFactory {
         CapabilitiesMiddleware preCapabilitiesMiddleware = CapabilitiesMiddleware.link(
                 new BrowserstackPreMiddleware(),
                 new SauceLabsPreMiddleware(),
-                new ZebrunnerPreMiddleware(testName));
+                new ZebrunnerPreMiddleware(testName),
+                new LambdaTestPreMiddleware());
 
         CapabilitiesMiddleware capabilitiesMiddleware = CapabilitiesMiddleware.link(
                 new ChromeCapabilitiesMiddleware(),
@@ -89,10 +93,12 @@ public class DriverFactory {
                 new XCUITestCapabilitiesMiddleware(),
                 new GeckoCapabilitiesMiddleware(),
                 new Mac2CapabilitiesMiddleware(),
-                new SafariCapabilitiesMiddleware());
+                new SafariCapabilitiesMiddleware(),
+                new WindowsCapabilitiesMiddleware());
 
         CapabilitiesMiddleware postCapabilitiesMiddleware = CapabilitiesMiddleware.link(
-                new BrowserstackPostMiddleware());
+                new BrowserstackPostMiddleware(),
+                new MCloudPostMiddleware());
 
         AbstractFactory driverFactory = AbstractFactory.link(
                 new ChromeFactory(),
